+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## Unreleased
+### Added:
+- `shuriken.iterator`.
+
 ## [0.14.51] - 2024-01-07
 ### Fixed:
 - Fix `monkey-patch`, `reflection` and `byte-buddy`.